+++ packages/frontend/core/src/blocksuite/ai/blocks/ai-chat-block/model/index.ts
@@ -1,3 +1,2 @@
 export * from './ai-chat-model';
 export * from './consts';
-export * from './types';
