+++ packages/frontend/core/src/blocksuite/ai/chat-panel/actions/image-to-text.ts
@@ -5,7 +5,7 @@ import { html, nothing } from 'lit';
 import { property } from 'lit/decorators.js';
 import { styleMap } from 'lit/directives/style-map.js';
 
-import type { ChatAction } from '../chat-context';
+import { type ChatAction } from '../../components/ai-chat-messages';
 
 export class ActionImageToText extends WithDisposable(ShadowlessElement) {
   @property({ attribute: false })
