+++ packages/frontend/core/src/blocksuite/ai/chat-panel/actions/image.ts
@@ -8,7 +8,7 @@ import { html, nothing } from 'lit';
 import { property } from 'lit/decorators.js';
 import { styleMap } from 'lit/directives/style-map.js';
 
-import type { ChatAction } from '../chat-context';
+import { type ChatAction } from '../../components/ai-chat-messages';
 
 export class ActionImage extends WithDisposable(ShadowlessElement) {
   @property({ attribute: false })
