+++ packages/frontend/core/src/blocksuite/ai/chat-panel/actions/mindmap.ts
@@ -7,7 +7,7 @@ import { html } from 'lit';
 import { property } from 'lit/decorators.js';
 import { styleMap } from 'lit/directives/style-map.js';
 
-import type { ChatAction } from '../chat-context';
+import { type ChatAction } from '../../components/ai-chat-messages';
 
 export class ActionMindmap extends WithDisposable(ShadowlessElement) {
   @property({ attribute: false })
