+++ packages/frontend/core/src/blocksuite/ai/chat-panel/const.ts
@@ -8,6 +8,3 @@ export const HISTORY_IMAGE_ACTIONS = [
   'Remove background',
   'Convert to sticker',
 ];
-
-export const PROMPT_NAME_AFFINE_AI = 'Chat With AFFiNE AI';
-export const PROMPT_NAME_NETWORK_SEARCH = 'Search With AFFiNE AI';
