+++ packages/frontend/core/src/blocksuite/ai/components/ai-chat-input/const.ts
@@ -1,0 +1,2 @@
+export const PROMPT_NAME_AFFINE_AI = 'Chat With AFFiNE AI';
+export const PROMPT_NAME_NETWORK_SEARCH = 'Search With AFFiNE AI';
