+++ packages/frontend/core/src/blocksuite/ai/components/ai-chat-input/index.ts
@@ -1,0 +1,3 @@
+export * from './ai-chat-input';
+export * from './const';
+export * from './type';
