+++ packages/frontend/core/src/blocksuite/ai/components/ai-chat-input/type.ts
@@ -1,0 +1,21 @@
+import type { Signal } from '@preact/signals-core';
+
+import type { AIError } from '../../provider';
+import type { ChatStatus, HistoryMessage } from '../ai-chat-messages';
+
+export interface AINetworkSearchConfig {
+  visible: Signal<boolean | undefined>;
+  enabled: Signal<boolean | undefined>;
+  setEnabled: (state: boolean) => void;
+}
+
+// TODO: remove this type
+export type AIChatInputContext = {
+  messages: HistoryMessage[];
+  status: ChatStatus;
+  error: AIError | null;
+  quote?: string;
+  markdown?: string;
+  images: File[];
+  abortController: AbortController | null;
+};
