+++ packages/frontend/core/src/blocksuite/ai/components/ai-chat-messages/index.ts
@@ -1,0 +1,1 @@
+export * from './type';
