+++ packages/frontend/core/src/blocksuite/ai/components/ai-chat-messages/type.ts
@@ -1,6 +1,5 @@
 import { z } from 'zod';
 
-// Define the Zod schema
 const ChatMessageSchema = z.object({
   id: z.string(),
   content: z.string(),
@@ -14,12 +13,36 @@ const ChatMessageSchema = z.object({
 
 export const ChatMessagesSchema = z.array(ChatMessageSchema);
 
-// Derive the TypeScript type from the Zod schema
 export type ChatMessage = z.infer<typeof ChatMessageSchema>;
 
+export type ChatAction = {
+  action: string;
+  messages: ChatMessage[];
+  sessionId: string;
+  createdAt: string;
+};
+
+export type HistoryMessage = ChatMessage | ChatAction;
+
 export type MessageRole = 'user' | 'assistant';
+
 export type MessageUserInfo = {
   userId?: string;
   userName?: string;
   avatarUrl?: string;
 };
+
+export function isChatAction(item: HistoryMessage): item is ChatAction {
+  return 'action' in item;
+}
+
+export function isChatMessage(item: HistoryMessage): item is ChatMessage {
+  return 'role' in item;
+}
+
+export type ChatStatus =
+  | 'loading'
+  | 'success'
+  | 'error'
+  | 'idle'
+  | 'transmitting';
