+++ packages/frontend/core/src/blocksuite/ai/peek-view/types.ts
@@ -1,12 +1,5 @@
-import type { ChatMessage } from '../blocks';
+import type { ChatMessage, ChatStatus } from '../components/ai-chat-messages';
 import type { AIError } from '../provider';
-
-export type ChatStatus =
-  | 'success'
-  | 'error'
-  | 'idle'
-  | 'transmitting'
-  | 'loading';
 
 export type ChatContext = {
   messages: ChatMessage[];
@@ -14,6 +7,4 @@ export type ChatContext = {
   error: AIError | null;
   images: File[];
   abortController: AbortController | null;
-  currentSessionId: string | null;
-  currentChatBlockId: string | null;
 };
